+++ src/mirrors/MirrorLoader.ts
@@ -54,6 +54,7 @@ import { MangaPark } from "./sites/MangaPark"
 import { MangaReaderTo } from "./sites/MangaReaderTo"
 import { FanComics } from "./sites/FanComics"
 import { MangaBTT } from "./sites/MangaBTT"
+import { MangaBuddy } from "./sites/MangaBuddy"
 
 export class MirrorLoader {
     lookupMap: Map<string, MirrorImplementation>
@@ -147,6 +148,7 @@ export const getMirrorLoader = (mirrorHelper: MirrorHelper) => {
             new MangaReaderTo(mirrorHelper),
             new FanComics(mirrorHelper),
             new MangaBTT(mirrorHelper),
+            new MangaBuddy(mirrorHelper),
 
             // Must be last, as we sometimes select based on same matching
             // domains (MangaDex), therefore we should prefer active ones first